fuck

+++ affinewarp/shiftwarp.py
@@ -38,7 +38,7 @@ class ShiftWarping(object):
         #   T = number of timepoints
         #   N = number of features/neurons
         K, T, N = data.shape
-        self.shifts = np.zeros(K)
+        self.shifts = np.zeros(K, dtype=int)
 
         L = int(self.maxlag * T)
         losses = np.empty((K, 2*L+1))
